A bit better Magnum DartApp:
